+++ e2e/__tests__/resolve_no_file_extensions.test.js
@@ -8,8 +8,14 @@
  */
 'use strict';
 
+import path from 'path';
 import runJest from '../runJest';
-import {extractSummary} from '../Utils';
+import {cleanup, extractSummary, writeFiles} from '../Utils';
+
+const DIR = path.resolve(__dirname, '../resolve_no_extensions-no-js');
+
+beforeEach(() => cleanup(DIR));
+afterAll(() => cleanup(DIR));
 
 test('show error message with matching files', () => {
   const {status, stderr} = runJest('resolve_no_extensions');
@@ -17,4 +23,31 @@ test('show error message with matching files', () => {
 
   expect(status).toBe(1);
   expect(rest).toMatchSnapshot();
+});
+
+test('show error message when no js moduleFileExtensions', () => {
+  writeFiles(DIR, {
+    'index.jsx': `
+      module.exports ={found: true};
+    `,
+    'package.json': `
+      {
+        "jest": {
+          "moduleFileExtensions": ["jsx"]
+        }
+      }
+    `,
+    'test.jsx': `
+      const m = require('../');
+
+      test('some test', () => {
+        expect(m.found).toBe(true);
+      });
+    `,
+  });
+
+  const {status, stderr} = runJest('resolve_no_extensions-no-js');
+
+  expect(status).toBe(1);
+  expect(stderr).toMatchSnapshot();
 });